Apply PR comments (extracted interfaces to separate files + small refactor)
