+++ src/functions/create.ts
@@ -1,7 +1,7 @@
 import { doc } from 'serverless-dynamodb-client';
 import { Context, Callback } from 'aws-lambda';
 
-import { ITestResult } from '../interfaces/interface';
+import { ITestResult } from '../interfaces/itest-result';
 import TestResultService from '../services/test-result';
 
 const testResultService = new TestResultService(doc, process.env.DYNAMODB_TABLE);
