+++ src/interfaces/iauth.ts
