+++ src/interfaces/icompetency.ts
@@ -1,0 +1,6 @@
+export interface ICompetency {
+	id: string,
+	faultsNo: number,
+	isSerious: boolean,
+	isDangerous: boolean
+}
