+++ src/interfaces/iresponse.ts
@@ -1,0 +1,5 @@
+export interface IResponse {
+	body: any,
+	statusCode: number,
+	headers: { [id: string]: string; }
+}
