+++ src/interfaces/itest-result.ts
@@ -1,0 +1,9 @@
+import { ICompetency } from "./icompetency";
+
+
+export interface ITestResult {
+	id?: string,
+	_candidateId: string,
+	faults: ICompetency[]
+}
+
