+++ src/services/no-auth.ts
@@ -1,4 +1,4 @@
-import { IAuth } from "./iauth";
+import { IAuth } from "../interfaces/iauth";
 
 /* eslint class-methods-use-this: "off" */
 /* eslint-env es6 */
